Split Menus, Repl, and UI modules
