+++ src/Swarm/TUI/Model.hs
@@ -110,11 +110,8 @@ module Swarm.TUI.Model (
 
   -- ** Initialization
   initFocusRing,
-  defaultPrompt,
-  initREPLState,
   initLgTicksPerSecond,
   initUIState,
-  lastEntry,
 
   -- ** Updating
   populateInventoryList,
@@ -151,26 +148,15 @@ module Swarm.TUI.Model (
 
 import Brick
 import Brick.Focus
-import Brick.Widgets.Dialog (Dialog)
-import Brick.Widgets.Edit (Editor, applyEdit, editorText, getEditContents)
 import Brick.Widgets.List qualified as BL
-import Control.Applicative (Applicative (liftA2))
 import Control.Lens hiding (from, (<.>))
 import Control.Monad.Except
 import Control.Monad.State
-import Data.Bits (FiniteBits (finiteBitSize))
-import Data.Foldable (toList)
 import Data.List (findIndex)
 import Data.List.NonEmpty (NonEmpty (..))
-import Data.List.NonEmpty qualified as NE
-import Data.Map (Map)
-import Data.Map qualified as M
-import Data.Maybe (fromMaybe, isJust)
-import Data.Sequence (Seq)
-import Data.Sequence qualified as Seq
+import Data.Maybe (fromMaybe)
 import Data.Text (Text)
 import Data.Text qualified as T
-import Data.Text.Zipper qualified as TZ
 import Data.Vector qualified as V
 import GitHash (GitInfo)
 import Graphics.Vty (ColorMode (..))
@@ -179,24 +165,18 @@ import Network.Wai.Handler.Warp (Port)
 import Swarm.Game.Entity as E
 import Swarm.Game.Robot
 import Swarm.Game.ScenarioInfo (
-  ScenarioCollection,
-  ScenarioInfo (..),
   ScenarioInfoPair,
-  ScenarioItem (..),
-  scMap,
-  scenarioCollectionToList,
   _SISingle,
  )
 import Swarm.Game.State
-import Swarm.Game.World qualified as W
-import Swarm.Language.Types
 import Swarm.TUI.Inventory.Sorting
+import Swarm.TUI.Model.Menus
 import Swarm.TUI.Model.Names
+import Swarm.TUI.Model.Repl
+import Swarm.TUI.Model.UI
 import Swarm.Util
 import Swarm.Version (NewReleaseFailure (NoMainUpstreamRelease))
 import System.Clock
-import System.FilePath (dropTrailingPathSeparator, splitPath, takeFileName)
-import Witch (into)
 
 ------------------------------------------------------------
 -- Custom UI label types
@@ -219,467 +199,6 @@ infoScroll = viewportScroll InfoViewport
 
 modalScroll :: ViewportScroll Name
 modalScroll = viewportScroll ModalViewport
-
-------------------------------------------------------------
--- REPL History
-------------------------------------------------------------
-
--- | An item in the REPL history.
-data REPLHistItem
-  = -- | Something entered by the user.
-    REPLEntry Text
-  | -- | A response printed by the system.
-    REPLOutput Text
-  deriving (Eq, Ord, Show, Read)
-
--- | Useful helper function to only get user input text.
-getREPLEntry :: REPLHistItem -> Maybe Text
-getREPLEntry = \case
-  REPLEntry t -> Just t
-  _ -> Nothing
-
--- | Useful helper function to filter out REPL output.
-isREPLEntry :: REPLHistItem -> Bool
-isREPLEntry = isJust . getREPLEntry
-
--- | Get the text of REPL input/output.
-replItemText :: REPLHistItem -> Text
-replItemText = \case
-  REPLEntry t -> t
-  REPLOutput t -> t
-
--- | History of the REPL with indices (0 is first entry) to the current
---   line and to the first entry since loading saved history.
---   We also (ab)use the length of the REPL as the index of current
---   input line, since that number is one past the index of last entry.
-data REPLHistory = REPLHistory
-  { _replSeq :: Seq REPLHistItem
-  , _replIndex :: Int
-  , _replStart :: Int
-  }
-
-makeLensesWith (lensRules & generateSignatures .~ False) ''REPLHistory
-
--- | Sequence of REPL inputs and outputs, oldest entry is leftmost.
-replSeq :: Lens' REPLHistory (Seq REPLHistItem)
-
--- | The current index in the REPL history (if the user is going back
---   through the history using up/down keys).
-replIndex :: Lens' REPLHistory Int
-
--- | The index of the first entry since loading saved history.
---
--- It will be set on load and reset on save (happens during exit).
-replStart :: Lens' REPLHistory Int
-
--- | Create new REPL history (i.e. from loaded history file lines).
-newREPLHistory :: [REPLHistItem] -> REPLHistory
-newREPLHistory xs =
-  let s = Seq.fromList xs
-   in REPLHistory
-        { _replSeq = s
-        , _replStart = length s
-        , _replIndex = length s
-        }
-
--- | Point the start of REPL history after current last line. See 'replStart'.
-restartREPLHistory :: REPLHistory -> REPLHistory
-restartREPLHistory h = h & replStart .~ replLength h
-
--- | Current number lines of the REPL history - (ab)used as index of input buffer.
-replLength :: REPLHistory -> Int
-replLength = length . _replSeq
-
--- | Add new REPL input - the index must have been pointing one past
---   the last element already, so we increment it to keep it that way.
-addREPLItem :: REPLHistItem -> REPLHistory -> REPLHistory
-addREPLItem t h =
-  h
-    & replSeq %~ (|> t)
-    & replIndex .~ 1 + replLength h
-
--- | Get the latest N items in history, starting with the oldest one.
---
--- This is used to show previous REPL lines in UI, so we need the items
--- sorted in the order they were entered and will be drawn top to bottom.
-getLatestREPLHistoryItems :: Int -> REPLHistory -> [REPLHistItem]
-getLatestREPLHistoryItems n h = toList latestN
- where
-  latestN = Seq.drop oldestIndex $ h ^. replSeq
-  oldestIndex = max (h ^. replStart) $ length (h ^. replSeq) - n
-
-data TimeDir = Newer | Older deriving (Eq, Ord, Show)
-
-moveReplHistIndex :: TimeDir -> Text -> REPLHistory -> REPLHistory
-moveReplHistIndex d lastEntered history = history & replIndex .~ newIndex
- where
-  historyLen = replLength history
-  curText = fromMaybe lastEntered $ getCurrentItemText history
-  curIndex = history ^. replIndex
-  entries = history ^. replSeq
-  -- split repl at index
-  (olderP, newer) = Seq.splitAt curIndex entries
-  -- find first different entry in direction
-  notSameEntry = \case
-    REPLEntry t -> t /= curText
-    _ -> False
-  newIndex = case d of
-    Newer -> maybe historyLen (curIndex +) $ Seq.findIndexL notSameEntry newer
-    Older -> fromMaybe curIndex $ Seq.findIndexR notSameEntry olderP
-
-getCurrentItemText :: REPLHistory -> Maybe Text
-getCurrentItemText history = replItemText <$> Seq.lookup (history ^. replIndex) (history ^. replSeq)
-
-replIndexIsAtInput :: REPLHistory -> Bool
-replIndexIsAtInput repl = repl ^. replIndex == replLength repl
-
--- | Given some text,  removes the REPLEntry within REPLHistory which is equal to that.
---   This is used when the user enters in search mode and want to traverse the history.
---   If a command has been used many times, the history will be populated with it causing
---   the effect that search command always finds the same command.
-removeEntry :: Text -> REPLHistory -> REPLHistory
-removeEntry foundtext hist = hist & replSeq %~ Seq.filter (/= REPLEntry foundtext)
-
--- | Get the last REPLEntry in REPLHistory matching the given text
-lastEntry :: Text -> REPLHistory -> Maybe Text
-lastEntry t h =
-  case Seq.viewr $ Seq.filter matchEntry $ h ^. replSeq of
-    Seq.EmptyR -> Nothing
-    _ Seq.:> a -> Just (replItemText a)
- where
-  matchesText histItem = t `T.isInfixOf` replItemText histItem
-  matchEntry = liftA2 (&&) matchesText isREPLEntry
-
-------------------------------------------------------------
--- REPL
-------------------------------------------------------------
-
--- | This data type tells us how to interpret the text typed
---   by the player at the prompt (which is stored in Editor).
-data REPLPrompt
-  = -- | Interpret the prompt text as a regular command.
-    --   The list is for potential completions, which we can
-    --   cycle through by hitting Tab repeatedly
-    CmdPrompt [Text]
-  | -- | Interpret the prompt text as "search this text in history"
-    SearchPrompt REPLHistory
-
-defaultPrompt :: REPLPrompt
-defaultPrompt = CmdPrompt []
-
-data ReplControlMode
-  = Piloting
-  | Typing
-  deriving (Enum, Bounded, Eq)
-
-data REPLState = REPLState
-  { _replPromptType :: REPLPrompt
-  , _replPromptEditor :: Editor Text Name
-  , _replValid :: Bool
-  , _replLast :: Text
-  , _replType :: Maybe Polytype
-  , _replControlMode :: ReplControlMode
-  , _replHistory :: REPLHistory
-  }
-
-newREPLEditor :: Text -> Editor Text Name
-newREPLEditor t = applyEdit gotoEnd $ editorText REPLInput (Just 1) t
- where
-  ls = T.lines t
-  pos = (length ls - 1, T.length (last ls))
-  gotoEnd = if null ls then id else TZ.moveCursor pos
-
-initREPLState :: REPLHistory -> REPLState
-initREPLState = REPLState defaultPrompt (newREPLEditor "") True "" Nothing Typing
-
-makeLensesWith (lensRules & generateSignatures .~ False) ''REPLState
-
--- | The way we interpret text typed by the player in the REPL prompt.
-replPromptType :: Lens' REPLState REPLPrompt
-
--- | The prompt where the user can type input at the REPL.
-replPromptEditor :: Lens' REPLState (Editor Text Name)
-
--- | Convinience lens to get text from editor and replace it with new
---   one that has the provided text.
-replPromptText :: Lens' REPLState Text
-replPromptText = lens g s
- where
-  g r = r ^. replPromptEditor . to getEditContents . to T.concat
-  s r t = r & replPromptEditor .~ newREPLEditor t
-
--- | Whether the prompt text is a valid 'Term'.
-replValid :: Lens' REPLState Bool
-
--- | The type of the current REPL input which should be displayed to
---   the user (if any).
-replType :: Lens' REPLState (Maybe Polytype)
-
--- | The last thing the user has typed which isn't part of the history.
---   This is used to restore the repl form after the user visited the history.
-replLast :: Lens' REPLState Text
-
--- | Piloting or Typing mode
-replControlMode :: Lens' REPLState ReplControlMode
-
--- | History of things the user has typed at the REPL, interleaved
---   with outputs the system has generated.
-replHistory :: Lens' REPLState REPLHistory
-
-------------------------------------------------------------
--- Menus and dialogs
-------------------------------------------------------------
-
-data ModalType
-  = HelpModal
-  | RecipesModal
-  | CommandsModal
-  | MessagesModal
-  | RobotsModal
-  | WinModal
-  | QuitModal
-  | KeepPlayingModal
-  | DescriptionModal Entity
-  | GoalModal [Text]
-  deriving (Eq, Show)
-
-data ButtonSelection = CancelButton | KeepPlayingButton | StartOverButton Seed ScenarioInfoPair | QuitButton | NextButton ScenarioInfoPair
-
-data Modal = Modal
-  { _modalType :: ModalType
-  , _modalDialog :: Dialog ButtonSelection
-  }
-
-makeLenses ''Modal
-
-data MainMenuEntry = NewGame | Tutorial | Messages | About | Quit
-  deriving (Eq, Ord, Show, Read, Bounded, Enum)
-
-data Menu
-  = NoMenu -- We started playing directly from command line, no menu to show
-  | MainMenu (BL.List Name MainMenuEntry)
-  | NewGameMenu (NonEmpty (BL.List Name ScenarioItem)) -- stack of scenario item lists
-  | MessagesMenu
-  | AboutMenu
-
-mainMenu :: MainMenuEntry -> BL.List Name MainMenuEntry
-mainMenu e = BL.list MenuList (V.fromList listEnums) 1 & BL.listMoveToElement e
-
-makePrisms ''Menu
-
--- | Create a brick 'BL.List' of scenario items from a 'ScenarioCollection'.
-mkScenarioList :: Bool -> ScenarioCollection -> BL.List Name ScenarioItem
-mkScenarioList cheat = flip (BL.list ScenarioList) 1 . V.fromList . filterTest . scenarioCollectionToList
- where
-  filterTest = if cheat then id else filter (\case SICollection n _ -> n /= "Testing"; _ -> True)
-
--- | Given a 'ScenarioCollection' and a 'FilePath' which is the canonical
---   path to some folder or scenario, construct a 'NewGameMenu' stack
---   focused on the given item, if possible.
-mkNewGameMenu :: Bool -> ScenarioCollection -> FilePath -> Maybe Menu
-mkNewGameMenu cheat sc path = NewGameMenu . NE.fromList <$> go (Just sc) (splitPath path) []
- where
-  go :: Maybe ScenarioCollection -> [FilePath] -> [BL.List Name ScenarioItem] -> Maybe [BL.List Name ScenarioItem]
-  go _ [] stk = Just stk
-  go Nothing _ _ = Nothing
-  go (Just curSC) (thing : rest) stk = go nextSC rest (lst : stk)
-   where
-    hasName :: ScenarioItem -> Bool
-    hasName (SISingle (_, ScenarioInfo pth _ _ _)) = takeFileName pth == thing
-    hasName (SICollection nm _) = nm == into @Text (dropTrailingPathSeparator thing)
-
-    lst = BL.listFindBy hasName (mkScenarioList cheat curSC)
-
-    nextSC = case M.lookup (dropTrailingPathSeparator thing) (scMap curSC) of
-      Just (SICollection _ c) -> Just c
-      _ -> Nothing
-
-------------------------------------------------------------
--- Inventory list entries
-------------------------------------------------------------
-
--- | An entry in the inventory list displayed in the info panel.  We
---   can either have an entity with a count in the robot's inventory,
---   an entity installed on the robot, or a labelled separator.  The
---   purpose of the separators is to show a clear distinction between
---   the robot's /inventory/ and its /installed devices/.
-data InventoryListEntry
-  = Separator Text
-  | InventoryEntry Count Entity
-  | InstalledEntry Entity
-  deriving (Eq)
-
-makePrisms ''InventoryListEntry
-
-------------------------------------------------------------
--- UI state
-------------------------------------------------------------
-
--- | The main record holding the UI state.  For access to the fields,
--- see the lenses below.
-data UIState = UIState
-  { _uiMenu :: Menu
-  , _uiPlaying :: Bool
-  , _uiCheatMode :: Bool
-  , _uiFocusRing :: FocusRing Name
-  , _uiWorldCursor :: Maybe W.Coords
-  , _uiREPL :: REPLState
-  , _uiInventory :: Maybe (Int, BL.List Name InventoryListEntry)
-  , _uiInventorySort :: InventorySortOptions
-  , _uiMoreInfoTop :: Bool
-  , _uiMoreInfoBot :: Bool
-  , _uiScrollToEnd :: Bool
-  , _uiError :: Maybe Text
-  , _uiModal :: Maybe Modal
-  , _uiGoal :: Maybe [Text]
-  , _uiShowFPS :: Bool
-  , _uiShowZero :: Bool
-  , _uiHideRobotsUntil :: TimeSpec
-  , _uiInventoryShouldUpdate :: Bool
-  , _uiTPF :: Double
-  , _uiFPS :: Double
-  , _lgTicksPerSecond :: Int
-  , _tickCount :: Int
-  , _frameCount :: Int
-  , _frameTickCount :: Int
-  , _lastFrameTime :: TimeSpec
-  , _accumulatedTime :: TimeSpec
-  , _lastInfoTime :: TimeSpec
-  , _appData :: Map Text Text
-  , _scenarioRef :: Maybe ScenarioInfoPair
-  }
-
---------------------------------------------------
--- Lenses for UIState
-
-let exclude = ['_lgTicksPerSecond]
- in makeLensesWith
-      ( lensRules
-          & generateSignatures .~ False
-          & lensField . mapped . mapped %~ \fn n ->
-            if n `elem` exclude then [] else fn n
-      )
-      ''UIState
-
--- | The current menu state.
-uiMenu :: Lens' UIState Menu
-
--- | Are we currently playing the game?  True = we are playing, and
---   should thus display a world, REPL, etc.; False = we should
---   display the current menu.
-uiPlaying :: Lens' UIState Bool
-
--- | Cheat mode, i.e. are we allowed to turn creative mode on and off?
-uiCheatMode :: Lens' UIState Bool
-
--- | The focus ring is the set of UI panels we can cycle among using
---   the Tab key.
-uiFocusRing :: Lens' UIState (FocusRing Name)
-
--- | The last clicked position on the world view.
-uiWorldCursor :: Lens' UIState (Maybe W.Coords)
-
--- | The state of REPL panel.
-uiREPL :: Lens' UIState REPLState
-
--- | The order and direction of sorting inventory list.
-uiInventorySort :: Lens' UIState InventorySortOptions
-
--- | The hash value of the focused robot entity (so we can tell if its
---   inventory changed) along with a list of the items in the
---   focused robot's inventory.
-uiInventory :: Lens' UIState (Maybe (Int, BL.List Name InventoryListEntry))
-
--- | Does the info panel contain more content past the top of the panel?
-uiMoreInfoTop :: Lens' UIState Bool
-
--- | Does the info panel contain more content past the bottom of the panel?
-uiMoreInfoBot :: Lens' UIState Bool
-
--- | A flag telling the UI to scroll the info panel to the very end
---   (used when a new log message is appended).
-uiScrollToEnd :: Lens' UIState Bool
-
--- | When this is @Just@, it represents a popup box containing an
---   error message that is shown on top of the rest of the UI.
-uiError :: Lens' UIState (Maybe Text)
-
--- | When this is @Just@, it represents a modal to be displayed on
---   top of the UI, e.g. for the Help screen.
-uiModal :: Lens' UIState (Maybe Modal)
-
--- | Status of the scenario goal: whether there is one, and whether it
---   has been displayed to the user initially.
-uiGoal :: Lens' UIState (Maybe [Text])
-
--- | A toggle to show the FPS by pressing `f`
-uiShowFPS :: Lens' UIState Bool
-
--- | A toggle to show or hide inventory items with count 0 by pressing `0`
-uiShowZero :: Lens' UIState Bool
-
--- | Hide robots on the world map.
-uiHideRobotsUntil :: Lens' UIState TimeSpec
-
--- | Whether to show or hide robots on the world map.
-uiShowRobots :: Getter UIState Bool
-uiShowRobots = to (\ui -> ui ^. lastFrameTime > ui ^. uiHideRobotsUntil)
-
--- | Whether the Inventory ui panel should update
-uiInventoryShouldUpdate :: Lens' UIState Bool
-
--- | Computed ticks per milli seconds
-uiTPF :: Lens' UIState Double
-
--- | Computed frames per milli seconds
-uiFPS :: Lens' UIState Double
-
--- | The currently active Scenario description, useful for starting over.
-scenarioRef :: Lens' UIState (Maybe ScenarioInfoPair)
-
--- | The base-2 logarithm of the current game speed in ticks/second.
---   Note that we cap this value to the range of +/- log2 INTMAX.
-lgTicksPerSecond :: Lens' UIState Int
-lgTicksPerSecond = lens _lgTicksPerSecond safeSetLgTicks
- where
-  maxLog = finiteBitSize (maxBound :: Int)
-  maxTicks = maxLog - 2
-  minTicks = 2 - maxLog
-  safeSetLgTicks ui lTicks
-    | lTicks < minTicks = setLgTicks ui minTicks
-    | lTicks > maxTicks = setLgTicks ui maxTicks
-    | otherwise = setLgTicks ui lTicks
-  setLgTicks ui lTicks = ui {_lgTicksPerSecond = lTicks}
-
--- | A counter used to track how many ticks have happened since the
---   last time we updated the ticks/frame statistics.
-tickCount :: Lens' UIState Int
-
--- | A counter used to track how many frames have been rendered since the
---   last time we updated the ticks/frame statistics.
-frameCount :: Lens' UIState Int
-
--- | A counter used to track how many ticks have happened in the
---   current frame, so we can stop when we get to the tick cap.
-frameTickCount :: Lens' UIState Int
-
--- | The time of the last info widget update
-lastInfoTime :: Lens' UIState TimeSpec
-
--- | The time of the last 'Frame' event.
-lastFrameTime :: Lens' UIState TimeSpec
-
--- | The amount of accumulated real time.  Every time we get a 'Frame'
---   event, we accumulate the amount of real time that happened since
---   the last frame, then attempt to take an appropriate number of
---   ticks to "catch up", based on the target tick rate.
---
---   See https://gafferongames.com/post/fix_your_timestep/ .
-accumulatedTime :: Lens' UIState TimeSpec
-
--- | Free-form data loaded from the @data@ directory, for things like
---   the logo, about page, tutorial story, etc.
-appData :: Lens' UIState (Map Text Text)
 
 -- ----------------------------------------------------------------------------
 --                                Runtime state                              --
