+++ src/Swarm/TUI/View.hs
@@ -96,6 +96,7 @@ import Swarm.TUI.Attr
 import Swarm.TUI.Border
 import Swarm.TUI.Inventory.Sorting (renderSortMethod)
 import Swarm.TUI.Model
+import Swarm.TUI.Model.Repl
 import Swarm.TUI.Panel
 import Swarm.Util
 import Swarm.Version (NewReleaseFailure (..))
