+++ swarm.cabal
@@ -114,8 +114,11 @@ library
                       Swarm.TUI.List
                       Swarm.TUI.Panel
                       Swarm.TUI.Model
+                      Swarm.TUI.Model.Menus
                       Swarm.TUI.Model.Names
+                      Swarm.TUI.Model.Repl
                       Swarm.TUI.Model.StateUpdates
+                      Swarm.TUI.Model.UI
                       Swarm.TUI.View
                       Swarm.TUI.Controller
                       Swarm.TUI.Inventory.Sorting
